refactor: src/apis/status 로 구조변경
