--- apps/client/src/apis/config/end-point.ts
@@ -1,1 +1,0 @@
-// @TODO 엔드포인트 정의
